apps/zero/backend/flask/dev/requirements.txt
requests
flask
pyngrok
flask_socketio
gunicorn
gevent
flask_sqlalchemy
flask-marshmallow
marshmallow-sqlalchemy
translatepy
flask_restful
flask_httpauth